src/data-kafka-service/requirements.txt
mimesis
kafka-python
